pyproject.toml
[project]
name = "records-request-download"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "nameparser>=1.1.3",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "selenium>=4.28.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.11.1",
    "ruff>=0.12.10",
]
